reconfig row and col order for board plays, TODOS for row
Reconfig row and col order to: new_board[row][col].... this is the proper order to choose a row and a col.

added TODO: count 1's in chosen col before assigning row its value. If 2 ones in a certain col then row should be set to 2 so that it chooses row 3 (0 indexed)