randomized column picking
break out of the else statement so that only one piece is placed in a column for a play, instead of filling up an entire column

Set row to 3 to have row start from bottom up, used subtraction to go up a column until an empty spot was found.

TODO: Get this to work with -1 and 1 before putting it into the main program flow. This would be a very good separate function to make and call.